+++ noa-harvester/noaharvester/utils.py
@@ -1,8 +1,13 @@
 """Utility functions"""
 import logging
 
+from kafka.errors import NoBrokersAvailable
+
 from pyproj import Transformer, CRS
 import shapefile
+
+from noaharvester.messaging.kafka_producer import KafkaProducer
+from noaharvester.messaging.message import Message
 
 logger = logging.getLogger(__name__)
 
@@ -56,3 +61,15 @@ def get_bbox_from_shp(shp_path: str, bbox_only: bool) -> list:
             bboxes.append([west, south, east, north])
 
     return bboxes
+
+
+def send_kafka_message(bootstrap_servers, topic, succeeded, failed):
+    schema_def = Message.schema_response()
+
+    try:
+        producer = KafkaProducer(bootstrap_servers=bootstrap_servers, schema=schema_def)
+        kafka_message = {"succeeded": succeeded, "failed": failed}
+        producer.send(topic=topic, key=None, value=kafka_message)
+    except NoBrokersAvailable as e:
+        logger.warning("No brokers available. Continuing without Kafka. Error: %s", e)
+        producer = None