+++ noa-stac-ingest/noastacingest/__init__.py
@@ -1,2 +1,2 @@
 # pylint:disable=missing-module-docstring
-__version__ = "0.11.2"
+__version__ = "0.3.0"